To resolve the resize issue in IE
In IE we are not able to resize the column. It keeps the original size only even after trying to drag and resize.
minColumnWidth var comes as NaN because of unsupported width property the way that has been used currently.
And becuase of this NaN value newColWidth also becomes NaN, and that effectively doesn't change width of column and keeps width as it is.